+++ lib/db_managers/calendar_db_manager/include/ICalendarDbManager.hpp
@@ -7,12 +7,12 @@
 
 class ICalendarDbManager {
 public:
-    virtual const Calendar& get(size_t) = 0;
-    virtual size_t add(const Calendar&) = 0;
+    virtual const Calendar& get(int) = 0;
+    virtual int add(const Calendar&) = 0;
     virtual void update(const Calendar&) = 0;
-    virtual void remove(size_t) = 0;
+    virtual void remove(int) = 0;
 
-    virtual std::vector<Event> getEvents(size_t) = 0;
+    virtual std::vector<Event> getEvents(int) = 0;
 
 };
