+++ lib/db_managers/comment_db_manager/include/CommentDbManagerMock.hpp
@@ -7,14 +7,14 @@ class CommentDbManagerMock : public ICommentDbManager {
 public:
 	CommentDbManagerMock();
 
-	const Comment& get(size_t comment_id);
-	size_t add(const Comment& comment);
+	const Comment& get(int comment_id);
+	int add(const Comment& comment);
 	void update(const Comment& comment);
-	void remove(size_t comment_id);
+	void remove(int comment_id);
 
 private:
 	std::vector<Comment> data_;
-	size_t aid_;
+	int aid_;
 
 };
