+++ lib/db_managers/comment_db_manager/src/CommentDbManagerMock.cpp
@@ -9,7 +9,7 @@ CommentDbManagerMock::CommentDbManagerMock() :
 	data_.emplace_back(0, 0, 0, "");
 }
 
-const Comment& CommentDbManagerMock::get(size_t comment_id) {
+const Comment& CommentDbManagerMock::get(int comment_id) {
 	for (auto e = data_.begin() + 1; e != data_.end(); e++)
 		if (e->id == comment_id)
 			return *e;
@@ -17,7 +17,7 @@ const Comment& CommentDbManagerMock::get(size_t comment_id) {
 	return data_[0];
 }
 
-size_t CommentDbManagerMock::add(const Comment& comment) {
+int CommentDbManagerMock::add(const Comment& comment) {
 	data_.emplace_back(
 		aid_,
 		comment.event_id,
@@ -34,7 +34,7 @@ void CommentDbManagerMock::update(const Comment& comment) {
 			e = comment;
 }
 
-void CommentDbManagerMock::remove(size_t comment_id) {
+void CommentDbManagerMock::remove(int comment_id) {
 	for (auto e = data_.begin() + 1; e != data_.end(); e++)
 		if (e->id == comment_id)
 			data_.erase(e);
