+++ lib/db_managers/directory_db_manager/src/DirectoryDbManagerMock.cpp
@@ -6,7 +6,7 @@ DirectoryDbManagerMock::DirectoryDbManagerMock(std::shared_ptr<DbMock> db) :
 	db_(db),
 	aid_(db_->directories.size()) {}
 
-const Directory& DirectoryDbManagerMock::get(size_t directory_id) {
+const Directory& DirectoryDbManagerMock::get(int directory_id) {
 	for (auto e = db_->directories.begin(); e != db_->directories.end(); e++) 
 		if (e->id == directory_id)
 			return *e;
@@ -14,7 +14,7 @@ const Directory& DirectoryDbManagerMock::get(size_t directory_id) {
 	return db_->directories[0];
 }
 
-size_t DirectoryDbManagerMock::add(const Directory& directory) {
+int DirectoryDbManagerMock::add(const Directory& directory) {
 	db_->directories.emplace_back(
 		aid_,
 		directory.node_id,
@@ -32,7 +32,7 @@ void DirectoryDbManagerMock::update(const Directory& directory) {
 			e = directory;
 }
 
-void DirectoryDbManagerMock::remove(size_t directory_id) {
+void DirectoryDbManagerMock::remove(int directory_id) {
 	for (auto e = db_->directories.begin() + 1; e != db_->directories.end(); e++)
 		if (e->id == directory_id)
 			db_->directories.erase(e);
