+++ lib/db_managers/event_db_manager/src/EventDbManagerMock.cpp
@@ -12,7 +12,7 @@ EventDbManagerMock::EventDbManagerMock(std::shared_ptr<DbMock> db) :
 	db_(db),
 	aid_(db->events.size()) {}
 
-const Event& EventDbManagerMock::get(size_t event_id) {
+const Event& EventDbManagerMock::get(int event_id) {
 	for (auto e = db_->events.begin() + 1; e != db_->events.end(); e++)
 		if (e->id == event_id)
 			return *e;
@@ -20,7 +20,7 @@ const Event& EventDbManagerMock::get(size_t event_id) {
 	return db_->events[0];
 }
 
-size_t EventDbManagerMock::add(const Event& event) {
+int EventDbManagerMock::add(const Event& event) {
 	db_->events.emplace_back(
 		aid_,
 		event.calendar_id,
@@ -39,13 +39,13 @@ void EventDbManagerMock::update(const Event& event) {
 			e = event;
 }
 
-void EventDbManagerMock::remove(size_t event_id) {
+void EventDbManagerMock::remove(int event_id) {
 	for (auto e = db_->events.begin() + 1; e != db_->events.end(); e++)
 		if (e->id == event_id)
 			db_->events.erase(e);
 }
 
-std::vector<Comment> EventDbManagerMock::getComments(size_t event_id) {
+std::vector<Comment> EventDbManagerMock::getComments(int event_id) {
 	std::vector<Comment> comments;
 
 	for (auto e = db_->comments.begin() + 1; e != db_->comments.end(); e++)
