+++ lib/db_managers/node_db_manager/include/INodeDbManager.hpp
@@ -6,11 +6,11 @@
 
 class INodeDbManager {
   public:
-    virtual const Node& get(size_t) = 0;
-    virtual size_t add(const Node&) = 0;
+    virtual const Node& get(int) = 0;
+    virtual int add(const Node&) = 0;
     virtual void update(const Node&) = 0;
-    virtual void remove(size_t) = 0;
+    virtual void remove(int) = 0;
 
-    virtual std::vector<Node> getChildren(size_t) = 0;
+    virtual std::vector<Node> getChildren(int) = 0;
 };
