+++ lib/db_managers/node_db_manager/include/NodeDbManagerMock.hpp
@@ -11,16 +11,16 @@ class NodeDbManagerMock : public INodeDbManager {
 public:
     NodeDbManagerMock(std::shared_ptr<DbMock> db);
 
-    const Node& get(size_t node_id);
-    size_t add(const Node& node);
+    const Node& get(int node_id);
+    int add(const Node& node);
     void update(const Node& node);
-    void remove(size_t node_id);
+    void remove(int node_id);
 
-    std::vector<Node> getChildren(size_t node_id);
+    std::vector<Node> getChildren(int node_id);
 
 private:
     std::shared_ptr<DbMock> db_;
-    size_t aid_;
+    int aid_;
 
 };
