+++ lib/db_managers/node_db_manager/src/NodeDbManagerMock.cpp
@@ -8,7 +8,7 @@ NodeDbManagerMock::NodeDbManagerMock(std::shared_ptr<DbMock> db) :
 	db_(db),
 	aid_(db_->nodes.size()) {}
 
-const Node& NodeDbManagerMock::get(size_t node_id) {
+const Node& NodeDbManagerMock::get(int node_id) {
 	for (auto e = db_->nodes.begin() + 1; e != db_->nodes.end(); e++)
 		if (e->id == node_id)
 			return *e;
@@ -16,7 +16,7 @@ const Node& NodeDbManagerMock::get(size_t node_id) {
 	return db_->nodes[0];
 }
 
-size_t NodeDbManagerMock::add(const Node& node) {
+int NodeDbManagerMock::add(const Node& node) {
 	db_->nodes.emplace_back(
 		aid_,
 		node.parent_id,
@@ -33,13 +33,13 @@ void NodeDbManagerMock::update(const Node& node) {
 			e = node;
 }
 
-void NodeDbManagerMock::remove(size_t node_id) {
+void NodeDbManagerMock::remove(int node_id) {
 	for (auto e = db_->nodes.begin() + 1; e != db_->nodes.end(); e++)
 		if (e->id == node_id)
 			db_->nodes.erase(e);
 }
 
-std::vector<Node> NodeDbManagerMock::getChildren(size_t node_id) {
+std::vector<Node> NodeDbManagerMock::getChildren(int node_id) {
 	std::vector<Node> children;
 
 	for (auto e : db_->nodes)
