+++ lib/db_managers/profile_db_manager/src/ProfileDbManagerMock.cpp
@@ -7,10 +7,10 @@ ProfileDbManagerMock::ProfileDbManagerMock() :
 	data_(),
 	aid_(1) {
 
-	data_.emplace_back(0, 0, 0, std::vector<size_t>());
+	data_.emplace_back(0, 0, 0, std::vector<int>());
 }
 
-const Profile& ProfileDbManagerMock::get(size_t profile_id) {
+const Profile& ProfileDbManagerMock::get(int profile_id) {
 	for (auto e = data_.begin() + 1; e != data_.end(); e++)
 		if (e->id == profile_id)
 			return *e;
@@ -18,7 +18,7 @@ const Profile& ProfileDbManagerMock::get(size_t profile_id) {
 	return data_[0];
 }
 
-size_t ProfileDbManagerMock::add(const Profile& profile) {
+int ProfileDbManagerMock::add(const Profile& profile) {
 	data_.emplace_back(
 		aid_,
 		profile.node_id,
@@ -35,7 +35,7 @@ void ProfileDbManagerMock::update(const Profile& profile) {
 			e = profile;
 }
 
-void ProfileDbManagerMock::remove(size_t profile_id) {
+void ProfileDbManagerMock::remove(int profile_id) {
 	for (auto e = data_.begin() + 1; e != data_.end(); e++)
 		if (e->id == profile_id)
 			data_.erase(e);
