+++ lib/db_managers/tag_db_manager/include/TagDbManagerMock.hpp
@@ -8,14 +8,14 @@ class TagDbManagerMock : public ITagDbManager {
 public:
 	TagDbManagerMock();
 
-	const Tag& get(size_t tag_id);
-	size_t add(const Tag& tag);
+	const Tag& get(int tag_id);
+	int add(const Tag& tag);
 	void update(const Tag& tag);
-	void remove(size_t tag_id);
+	void remove(int tag_id);
 	
 private:
 	std::vector<Tag> data_;
-	size_t aid_;
+	int aid_;
 
 };
