+++ lib/db_managers/tag_db_manager/src/TagDbManagerMock.cpp
@@ -7,7 +7,7 @@ TagDbManagerMock::TagDbManagerMock() :
 	data_.emplace_back(0, "");
 }
 
-const Tag& TagDbManagerMock::get(size_t tag_id) {
+const Tag& TagDbManagerMock::get(int tag_id) {
 	for (auto e = data_.begin() + 1; e != data_.end(); e++)
 		if (e->id == tag_id)
 			return *e;
@@ -15,7 +15,7 @@ const Tag& TagDbManagerMock::get(size_t tag_id) {
 	return data_[0];
 }
 
-size_t TagDbManagerMock::add(const Tag& tag) {
+int TagDbManagerMock::add(const Tag& tag) {
 	data_.emplace_back(
 		aid_,
 		tag.name
@@ -30,7 +30,7 @@ void TagDbManagerMock::update(const Tag& tag) {
 			e = tag;
 }
 
-void TagDbManagerMock::remove(size_t tag_id) {
+void TagDbManagerMock::remove(int tag_id) {
 	for (auto e = data_.begin() + 1; e != data_.end(); e++)
 		if (e->id == tag_id)
 			data_.erase(e);
