+++ lib/db_managers/user_db_manager/include/IUserDbManager.hpp
@@ -5,7 +5,7 @@
 class IUserDbManager {
 public:
     virtual const User& get() = 0;
-    virtual const User& get(size_t user_id) = 0;
+    virtual const User& get(int user_id) = 0;
 
 };
