+++ lib/db_managers/user_db_manager/src/UserDbManagerMock.cpp
@@ -8,7 +8,7 @@ const User& UserDbManagerMock::get() {
 	return db_->users[0];
 }
 
-const User& UserDbManagerMock::get(size_t user_id) {
+const User& UserDbManagerMock::get(int user_id) {
 	return db_->users[user_id];
 }
