+++ lib/entities/calendar/include/Calendar.hpp
@@ -7,16 +7,16 @@ class Calendar {
   public:
     Calendar();
     Calendar(
-        size_t id,
-        size_t node_id,
-        size_t owner_id,
+        int id,
+        int node_id,
+        int owner_id,
         std::string name,
         std::string description
     );
 
-    size_t id;
-    size_t node_id;
-    size_t owner_id;
+    int id;
+    int node_id;
+    int owner_id;
 
     std::string name;
     std::string description;
