+++ lib/entities/calendar/src/Calendar.cpp
@@ -8,9 +8,9 @@ Calendar::Calendar() :
 	description() {}
 
 Calendar::Calendar(
-    size_t id,
-    size_t node_id,
-    size_t owner_id,
+    int id,
+    int node_id,
+    int owner_id,
     std::string name,
     std::string description
 ) :
