+++ lib/entities/comment/include/Comment.hpp
@@ -6,11 +6,11 @@
 class Comment {
 public:
 	Comment();
-	Comment(size_t id, size_t event_id, size_t owner_id, std::string text);
+	Comment(int id, int event_id, int owner_id, std::string text);
 
-	size_t id;
-	size_t event_id;
-	size_t owner_id;
+	int id;
+	int event_id;
+	int owner_id;
 
 	std::string text;
