+++ lib/entities/comment/src/Comment.cpp
@@ -8,7 +8,7 @@ Comment::Comment() :
 	owner_id(),
 	text() {}
 
-Comment::Comment(size_t id, size_t event_id, size_t owner_id, std::string text) :
+Comment::Comment(int id, int event_id, int owner_id, std::string text) :
 	id(id),
 	event_id(event_id),
 	owner_id(owner_id),
