+++ lib/entities/directory/src/Directory.cpp
@@ -7,7 +7,7 @@ Directory::Directory() :
 	name(),
 	description() {}
 
-Directory::Directory(size_t id, size_t node_id, size_t owner_id, std::string name, std::string description) :
+Directory::Directory(int id, int node_id, int owner_id, std::string name, std::string description) :
 	id(id),
 	node_id(node_id),
 	owner_id(owner_id),
