+++ lib/entities/event/include/Event.hpp
@@ -8,16 +8,16 @@ class Event {
 public:
     Event();
     Event(
-        size_t id,
-        size_t calendar_id,
+        int id,
+        int calendar_id,
         std::string name,
         std::string description,
         std::chrono::time_point<std::chrono::steady_clock> begin_point,
         std::chrono::time_point<std::chrono::steady_clock> end_point
     );
 
-    size_t id;
-    size_t calendar_id;
+    int id;
+    int calendar_id;
     std::string name;
     std::string description;
     std::chrono::time_point<std::chrono::steady_clock> begin_point;
