+++ lib/entities/event/src/Event.cpp
@@ -9,8 +9,8 @@ Event::Event() :
 	end_point() {}
 
 Event::Event(
-    size_t id,
-    size_t calendar_id,
+    int id,
+    int calendar_id,
     std::string name,
     std::string description,
     std::chrono::time_point<std::chrono::steady_clock> begin_point,
