+++ lib/entities/profile/include/Profile.hpp
@@ -6,13 +6,13 @@
 class Profile {
 public:
     Profile();
-    Profile(size_t id, size_t node_id, size_t owner_id, std::vector<size_t> nodes);
+    Profile(int id, int node_id, int owner_id, std::vector<int> nodes);
 
-    size_t id;
-    size_t node_id;
-    size_t owner_id;
+    int id;
+    int node_id;
+    int owner_id;
 
-    std::vector<size_t> nodes;
+    std::vector<int> nodes;
 
 };
