+++ lib/entities/profile/src/Profile.cpp
@@ -7,7 +7,7 @@ Profile::Profile() :
 	nodes() {}
 
 
-Profile::Profile(size_t id, size_t node_id, size_t owner_id, std::vector<size_t> nodes) :
+Profile::Profile(int id, int node_id, int owner_id, std::vector<int> nodes) :
 	id(id),
 	node_id(node_id),
 	owner_id(owner_id),
