+++ lib/entities/tag/src/Tag.cpp
@@ -4,7 +4,7 @@ Tag::Tag() :
 	id(),
 	name() {}
 
-Tag::Tag(size_t id, std::string name) :
+Tag::Tag(int id, std::string name) :
 	id(id),
 	name(name) {}
