+++ lib/entities/user/src/User.cpp
@@ -8,7 +8,7 @@ User::User() :
 	nickname(),
 	description() {}
 
-User::User(size_t id, size_t root_id, std::string login, std::string email, std::string nickname, std::string description) :
+User::User(int id, int root_id, std::string login, std::string email, std::string nickname, std::string description) :
 	id(id),
 	root_id(root_id),
 	login(login),
