+++ lib/managers/calendar_manager/src/CalendarManager.cpp
@@ -9,11 +9,11 @@
 CalendarManager::CalendarManager(std::shared_ptr<IDbManagers> db) :
     db_(db) {}
 
-const Calendar& CalendarManager::get(size_t calendar_id) {
+const Calendar& CalendarManager::get(int calendar_id) {
     return db_->calendar_dbm()->get(calendar_id);
 }
 
-size_t CalendarManager::add(const Calendar& calendar) {
+int CalendarManager::add(const Calendar& calendar) {
     return db_->calendar_dbm()->add(calendar);
 }
 
@@ -21,11 +21,11 @@ void CalendarManager::update(const Calendar& calendar) {
     db_->calendar_dbm()->add(calendar);
 }
 
-void CalendarManager::remove(size_t calendar_id) {
+void CalendarManager::remove(int calendar_id) {
     db_->calendar_dbm()->remove(calendar_id);
 }
 
-std::vector<Event> CalendarManager::getEvents(size_t calendar_id) {
+std::vector<Event> CalendarManager::getEvents(int calendar_id) {
     return db_->calendar_dbm()->getEvents(calendar_id);
 }
