+++ lib/managers/event_manager/include/EventManager.hpp
@@ -14,12 +14,12 @@ public:
 
     EventManager(std::shared_ptr<IDbManagers> db);
 
-    const Event& get(size_t event_id) override;
-    size_t add(const Event& event) override;
+    const Event& get(int event_id) override;
+    int add(const Event& event) override;
     void update(const Event& event) override;
-    void remove(size_t event_id) override;
+    void remove(int event_id) override;
 
-    std::vector<Comment> getComments(size_t event_id) override;
+    std::vector<Comment> getComments(int event_id) override;
     std::chrono::time_point<clock_t> parseTime(std::string str_time) override;
 
 private:
