+++ lib/managers/node_manager/include/NodeManager.hpp
@@ -13,18 +13,18 @@ class NodeManager : public INodeManager {
 public:
 	NodeManager(std::shared_ptr<IDbManagers> db);
 
-	const Node& get(size_t node_id) override;
-	size_t add(const Node& node) override;
+	const Node& get(int node_id) override;
+	int add(const Node& node) override;
 	void update(const Node& node) override;
 
 	// Рекурсивное удаление ноды и связанных с ней ресурсов
-	void remove(size_t node_id) override;
+	void remove(int node_id) override;
 
-	void tag(const Tag& tag, size_t node_id) override;
-	void move(size_t node_id, size_t destination_id) override;
-	void subscribe(size_t node_id) override;
-	void unsubscribe(size_t node_id) override;
-	std::vector<Node> getChildren(size_t node_id) override;
+	void tag(const Tag& tag, int node_id) override;
+	void move(int node_id, int destination_id) override;
+	void subscribe(int node_id) override;
+	void unsubscribe(int node_id) override;
+	std::vector<Node> getChildren(int node_id) override;
 
 private:
 	std::shared_ptr<IDbManagers> db_;
