+++ lib/managers/user_manager/include/IUserManager.hpp
@@ -5,7 +5,7 @@
 class IUserManager {
 public:
 	virtual const User& get() = 0;
-	virtual const User& get(size_t) = 0;
+	virtual const User& get(int) = 0;
 
 };
