+++ lib/managers/user_manager/include/UserManager.hpp
@@ -11,7 +11,7 @@ public:
 	UserManager(std::shared_ptr<IDbManagers> db);
 	
     const User& get();
-    const User& get(size_t user_id);
+    const User& get(int user_id);
 
 private:
 	std::shared_ptr<IDbManagers> db_;
