+++ lib/utils/tree/src/Tree.cpp
@@ -73,6 +73,6 @@ void Tree::uncheckNode(ITreeNode* node) {
     }
 }
 
-size_t Tree::checked() {
+int Tree::checked() {
     return checked_;
 }
