+++ lib/utils/tree/src/TreeNode.cpp
@@ -29,7 +29,7 @@ ITreeNode* TreeNode::getParent() { return parent_; }
 std::vector<ITreeNode*> TreeNode::getChildren() {
     std::vector<ITreeNode*> children;
 
-    for (size_t i = 0; i < children_.size(); i++)
+    for (int i = 0; i < children_.size(); i++)
         children.push_back(children_[i].get());
 
     return children;
